renamed the rest of the analysis files
