+++ src/itertree/examples/performance_analysis/check_get_by_idx.py
