+++ src/itertree/examples/performance_analysis/check_insert.py
