+++ src/itertree/examples/performance_analysis/check_iter.py
